Group request generators in requestgenerators module
